Move blazar core to init module
